requirements.txt
aiogram>=3.3.0
python-dotenv
rapidfuzz==3.9.6
numpy
Pillow
opencv-python-headless
pytesseract
easyocr
torch
torchvision
